📐 Architect: [Refactor/Health improvement] Split mcReducer
